change motor parameter to match what's shipped in the Filastruder kit
I'm not sure if the goal is to support exactly what's in the Filastruder
kit, but, the stepping angle doesn't match what's in the config.g file.

If there's a different machine build that this is supposed to target,
please ignore this PR, but maybe it would be helpful to have a README
or something like that somewhere which notes the deviation?